pyproject.toml
[tool.poetry]
name = "test"
version = "0.1.0"
description = ""
authors = ["aryadovoy <ryav.omsk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "0.19.0"
alembic = "1.8.1"
anyio = "3.6.1"
asgiref = "3.5.2"
asyncpg = "0.26.0"
backports-cached-property = "1.0.2"
bcrypt = "4.0.0"
boto3 = "1.24.86"
botocore = "1.27.86"
cffi = "1.15.1"
click = "8.1.3"
configargparse = "1.5.3"
cryptography = "38.0.1"
dnspython = "2.2.1"
ecdsa = "0.18.0"
email-validator = "1.3.0"
fastapi = "0.87.0"
graphql-core = "3.2.3"
greenlet = "1.1.3"
h11 = "0.12.0"
httptools = "0.5.0"
idna = "3.4"
importlib-metadata = "5.0.0"
importlib-resources = "5.9.0"
itsdangerous = "2.1.2"
jinja2 = "3.1.2"
makefun = "1.15.0"
mako = "1.2.3"
markupsafe = "2.1.1"
passlib = "1.7.4"
psycopg2-binary = "2.9.5"
pyasn1 = "0.4.8"
pycodestyle = "2.9.1"
pycparser = "2.21"
pydantic = "1.10.2"
pygments = "2.13.0"
pyjwt = "2.5.0"
pytest = "7.1.3"
python-dateutil = "2.8.2"
python-dotenv = "0.21.0"
python-jose = "3.3.0"
python-multipart = "0.0.5"
pyyaml = "6.0"
rsa = "4.9"
six = "1.16.0"
sniffio = "1.3.0"
sqladmin = "0.6.1"
sqlalchemy = "1.4.41"
sqlalchemy-utils = "0.38.3"
starlette = "0.21.0"
strawberry-graphql = "0.142.1"
toml = "0.10.2"
typing-extensions = "4.3.0"
uvloop = "0.17.0"
watchgod = "0.8.2"
websockets = "10.3"
wtforms = "3.0.1"
yarl = "1.8.1"
zipp = "3.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
